Removed session expiration checking from useQuery wrapper and react-query wrappers refactor
