+++ src/api/react-query-wrappers/index.ts
@@ -1,0 +1,2 @@
+export { useMutationWrapper } from './useMutationWrapper'
+export { useQueryWrapper } from './useQueryWrapper'
