+++ src/api/react-query-wrappers/useMutationWrapper.ts
@@ -1,35 +1,88 @@
 import React from 'react'
 import { useDialog, useLoadingOverlay, useToastNotification } from '@/contexts'
-import { useMutation, useQuery } from '@tanstack/react-query'
-import { UseMutationWrapper, UseQueryWrapper } from './comon.api.types'
+import { useMutation } from '@tanstack/react-query'
+import { UseMutationWrapper } from '../comon.api.types'
 import { useJwt } from '@/hooks/useJwt'
-import { NotFoundError } from '@/utils/errors.utils'
 
-export const useQueryWrapper: UseQueryWrapper = (key, queryFn, options) => {
-  const { hasSessionExpired, jwt } = useJwt()
-  const { openDialog } = useDialog()
-
-  if (!options?.skipJwtSessionExpirationCheck && hasSessionExpired()) {
-    openDialog({ type: 'sessionExpired' })
-  }
-
-  return useQuery(key, queryFn, {
-    useErrorBoundary: (error) => {
-      if (options?.skipThrowOn404Error && error instanceof NotFoundError) {
-        return false
-      }
-      return true
-    },
-    enabled: !!jwt && (options?.enabled ?? true),
-    ...options,
-  })
-}
-
+/**
+ * This abstract and takes care of most of what's needed for the application mutations:
+ * toast notifications, loading overlay, confirmation dialog and session expiration checks.
+ *
+ * It adds the following options:
+ *
+ * - `successToastLabel` - Required only if `suppressSuccessToast` is `false` or `undefined`.
+ * - `suppressSuccessToast` - If `true`, does not show success toast.
+ * - `errorToastLabel` - Required only if `suppressErrorToast` is `false` or `undefined`.
+ * - `suppressErrorToast` - If `true`, does not show error toast.
+ * - `loadingOverlayLabel` - Required only if `suppressLoadingOverlay` is `false` or `undefined`.
+ * - `suppressLoadingOverlay` - If `true`, does not show loading overlay.
+ * - `showConfirmationDialog` - If `true` shows the confirmation dialog when the mutation function is called.
+ * - `dialogConfig` - Options for dialog's title, description and button proceed label.
+ *
+ * On each label property, both function or string can be passed.
+ * The function will receive as a parameter the object passed to the mutate function,
+ * and only for the `successToastLabel` property, also the mutation's result data.
+ *
+ * @example
+ * const { mutate } = useMutationWrapper(..., {
+ *   ...,
+ *   successToastLabel: (data, variables) => {
+ *     console.log(data) // The data returned from the server on mutation call
+ *     console.log(variables) // The variables passed to the mutation function call
+ *     return t(`mutation-feedback.success`, { name: variables.eserviceName })
+ *   },
+ *   errorToastLabel: "ERROR",
+ *   loadingOverlay: (variables) => {
+ *     return t(`mutation-feedback.loading`, { name: variables.eserviceName })
+ *   },
+ * })
+ * ...
+ * mutate({ ..., eserviceName: "TEST" })
+ *
+ */
 export const useMutationWrapper: UseMutationWrapper = (mutationFn, options) => {
   const { showToast } = useToastNotification()
   const { showOverlay, hideOverlay } = useLoadingOverlay()
   const { openDialog, closeDialog } = useDialog()
   const { hasSessionExpired } = useJwt()
+
+  /**
+   * Wraps the react-query's onError option property. Handles the success toast notification.
+   */
+  const _wrappedOnSuccess: typeof options.onSuccess = (...args) => {
+    if (!options?.suppressSuccessToast && options?.successToastLabel) {
+      const successLabel =
+        typeof options.successToastLabel === 'function'
+          ? options.successToastLabel(...args)
+          : options?.successToastLabel
+
+      showToast(successLabel, 'success')
+    }
+    options?.onSuccess?.(...args)
+  }
+
+  /**
+   * Wraps the react-query's onError option property. Handles the error toast notification.
+   */
+  const _wrappedOnError: typeof options.onError = (...args) => {
+    if (!options?.suppressErrorToast && options.errorToastLabel) {
+      const errorLabel =
+        typeof options.errorToastLabel === 'function'
+          ? options.errorToastLabel(args[1], args[2])
+          : options?.errorToastLabel
+
+      showToast(errorLabel, 'error')
+    }
+    options?.onError?.(...args)
+  }
+
+  /**
+   * Wraps the react-query's onSettled option property. Hides the loading overlay.
+   */
+  const _wrappedOnSettled: typeof options.onSettled = (...args) => {
+    hideOverlay()
+    options?.onSettled?.(...args)
+  }
 
   const {
     mutate: _mutate,
@@ -37,34 +90,14 @@ export const useMutationWrapper: UseMutationWrapper = (mutationFn, options) => {
     ...rest
   } = useMutation(mutationFn, {
     ...options,
-    onSuccess: (...args) => {
-      if (!options?.suppressSuccessToast && options?.successToastLabel) {
-        const successLabel =
-          typeof options.successToastLabel === 'function'
-            ? options.successToastLabel(...args)
-            : options?.successToastLabel
-
-        showToast(successLabel, 'success')
-      }
-      options?.onSuccess?.(...args)
-    },
-    onError: (...args) => {
-      if (!options?.suppressErrorToast && options.errorToastLabel) {
-        const errorLabel =
-          typeof options.errorToastLabel === 'function'
-            ? options.errorToastLabel(args[1], args[2])
-            : options?.errorToastLabel
-
-        showToast(errorLabel, 'error')
-      }
-      options?.onError?.(...args)
-    },
-    onSettled: (...args) => {
-      hideOverlay()
-      options?.onSettled?.(...args)
-    },
+    onSuccess: _wrappedOnSuccess,
+    onError: _wrappedOnError,
+    onSettled: _wrappedOnSettled,
   })
 
+  /**
+   * This function
+   */
   const _wrapActionInDialog = React.useCallback(
     async <T>(
       action: (...args: unknown[]) => T,
